Integrate ANS + noise-aware compression into encoder/decoder pipeline
Encoder:
- Phase 0.5: pre-transform noise estimation + adaptive quantization
- Phase 1: parallel wavelet transform (4 threads)
- Phase 1.8: parallel ANS pre-encoding (4 threads)
- Phase 2: serial bitstream with per-band VLC/ANS auto-selection
- Negative quant sentinel for 16-bit skip-uncompand path

Decoder:
- ANS mode dispatch (modes 1-4) with jans_decode_band_x4
- Negative quant → skip uncompanding in dequantization
- Component clamping for Q6-Q8 overflow prevention
- NEON-accelerated dequantization paths
- Production hardening: assert(0) → proper error returns

+++ source/lib/vc5_encoder/syntax.c
@@ -106,11 +106,11 @@ CODEC_ERROR PutTagPair(BITSTREAM *stream, int tag, int value)
     // The bitstream should be aligned on a tag word boundary
     assert(IsAlignedTag(stream));
     
-    // The value must fit within a tag word
-    assert(((uint32_t)value & ~(uint32_t)CODEC_TAG_MASK) == 0);
-    
-    PutLong(stream, ((uint32_t )tag << 16) | (value & CODEC_TAG_MASK));
-    
+    // The value must fit within a tag word (mask to 16 bits to handle sign-extended TAGWORD values)
+    assert(((uint32_t)(uint16_t)value & ~(uint32_t)CODEC_TAG_MASK) == 0);
+
+    PutLong(stream, ((uint32_t )tag << 16) | ((uint32_t)(uint16_t)value & CODEC_TAG_MASK));
+
     return CODEC_ERROR_OKAY;
 }
 
@@ -123,16 +123,16 @@ CODEC_ERROR PutTagPairOptional(BITSTREAM *stream, int tag, int value)
 {
     // The bitstream should be aligned on a tag word boundary
     assert(IsAlignedTag(stream));
-    
-    // The value must fit within a tag word
-    assert(((uint32_t)value & ~(uint32_t)CODEC_TAG_MASK) == 0);
-    
+
+    // The value must fit within a tag word (mask to 16 bits to handle sign-extended TAGWORD values)
+    assert(((uint32_t)(uint16_t)value & ~(uint32_t)CODEC_TAG_MASK) == 0);
+
     // Set the optional tag bit
     //tag |= CODEC_TAG_OPTIONAL;
     //tag = NEG(tag);
     tag = neg(tag);
-    
-    PutLong(stream, ((uint32_t )tag << 16) | (value & CODEC_TAG_MASK));
+
+    PutLong(stream, ((uint32_t )tag << 16) | ((uint32_t)(uint16_t)value & CODEC_TAG_MASK));
     
     return CODEC_ERROR_OKAY;
 }
